rename game to golflike
