+++ golflike/common.lua
