+++ golflike/draw.lua
@@ -1,8 +1,8 @@
 --- draw.lua
 -- Handling of terminal I/O
-local common      = require("game.common")
-local competition = require('game.competition')
-local map         = require("game.map")
+local common      = require("golflike.common")
+local competition = require('golflike.competition')
+local map         = require("golflike.map")
 local termio      = require("term.io")
 local colour      = require('term.colour')
