+++ golflike/help.lua
@@ -2,10 +2,10 @@
 -- Help screen, displays a map key, controls and club ranges.
 local class  = require('30log')
 local cell   = require('forma.cell')
-local clubs  = require('game.clubs')
-local map    = require('game.map')
-local draw   = require('game.draw')
-local keymap = require('game.keymap')
+local clubs  = require('golflike.clubs')
+local map    = require('golflike.map')
+local draw   = require('golflike.draw')
+local keymap = require('golflike.keymap')
 local termio = require('term.io')
 local keys   = require('term.keys')
 local colour = require('term.colour')
