+++ golflike/mapgen.lua
@@ -4,7 +4,7 @@ local neighbourhood = require('forma.neighbourhood')
 local subpattern    = require('forma.subpattern')
 local automata      = require('forma.automata')
 local cell          = require('forma.cell')
-local utl           = require('game.util')
+local utl           = require('golflike.util')
 
 -- Performs a uniform sampling of the domain
 local function uniform(seedfrac)
