+++ golflike/markov.lua
