+++ golflike/util.lua
