+++ main.lua
@@ -1,11 +1,11 @@
 -- main.lua
 -- Game loop
-local log         = require('game.lib.log')
-local common      = require('game.common')
-local competition = require('game.competition')
-local state       = require('game.state')
-local maprender   = require('game.maprender')
-local menu        = require('game.menu')
+local log         = require('golflike.lib.log')
+local common      = require('golflike.common')
+local competition = require('golflike.competition')
+local state       = require('golflike.state')
+local maprender   = require('golflike.maprender')
+local menu        = require('golflike.menu')
 local termio      = require('term.io')
 
 local game = {}
